refactor: membership 폴더를 assets로 이동
