+++ src/assets/membership/가입완료버튼/complete_hover.png
